fix: check github username in middleware and pass it to signer object afterwards
