+++ fplus-lib/src/core/mod.rs
@@ -44,8 +44,15 @@ pub struct CreateApplicationInfo {
 pub struct VerifierList(pub Vec<String>);
 
 #[derive(Deserialize, Serialize, Debug)]
+pub struct ApplicationProposalApprovalSignerInfo {
+    pub signing_address: String,
+    pub created_at: String,
+    pub message_cid: String,
+}
+
+#[derive(Deserialize, Serialize, Debug)]
 pub struct CompleteNewApplicationProposalInfo {
-    pub signer: VerifierInput,
+    pub signer: ApplicationProposalApprovalSignerInfo,
     pub request_id: String,
 }
 
@@ -122,6 +129,14 @@ pub struct GithubQueryParams {
 
 #[derive(Deserialize)]
 pub struct ApplicationQueryParams {
+    pub id: String,
+    pub owner: String,
+    pub repo: String,
+}
+
+#[derive(Deserialize)]
+pub struct VerifierActionsQueryParams {
+    pub github_username: String,
     pub id: String,
     pub owner: String,
     pub repo: String,
